backend/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain
groq
mysql-connector-python
python-dotenv